Changed client side handlers for created, updated, destroyed, and reset to work with naked models
When using the syncing via this.signalRHub.syncUpdates(this); in a
model not a collection the client handlers would fail because models
dont have add, remove or reset methods. I added some conditional logic
to check if those methods were on the collection object and if not
treat it as a model. The only one I’m not sure of is the destroy
method, I just made the model=null, but I dont know if this is the
correct way to destroy a BB model clientside.